+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixes a bug when displaying an insurance and it doesn't have an address
 - Fixes a bug redirecting after tracker create
+- Fixes a bug related to required params on insurance create calls
 
 ## v0.13.0 (2024-08-27)